requirements.txt
aiohttp
bs4
selenium
requests
mechanicalsoup
faker
loguru
grpcio
grpcio-tools